Refactoring, creating a new constant class
