+++ spring-boot/src/main/java/packages/backendlogic/LoanTableConstant.java
@@ -1,0 +1,45 @@
+package packages.backendlogic;
+
+public class LoanTableConstant {
+    public final int CREDIT_SCORE_RANGE1_MIN = 300;
+    public final int CREDIT_SCORE_RANGE1_MAX = 500;
+    public final int CREDIT_SCORE_RANGE2_MIN = 501;
+    public final int CREDIT_SCORE_RANGE2_MAX = 600;
+    public final int CREDIT_SCORE_RANGE3_MIN = 601;
+    public final int CREDIT_SCORE_RANGE3_MAX = 660;
+    public final int CREDIT_SCORE_RANGE4_MIN = 661;
+    public final int CREDIT_SCORE_RANGE4_MAX = 700;
+
+    public final double PTI_RANGE1_MIN = 0.16;
+    public final double PTI_RANGE1_MAX = 0.2;
+    public final double PTI_RANGE2_MIN = 0.13;
+    public final double PTI_RANGE2_MAX = 0.15;
+    public final double PTI_RANGE3_MIN = 0.11;
+    public final double PTI_RANGE3_MAX = 0.12;
+    public final double PTI_RANGE4_MIN = 0.09;
+    public final double PTI_RANGE4_MAX = 0.1;
+
+    public final double DTI_RANGE1 = 0.5;
+    public final double DTI_RANGE2_MIN = 0.4;
+    public final double DTI_RANGE2_MAX = 0.49;
+    public final double DTI_RANGE3_MIN = 0.36;
+    public final double DTI_RANGE3_MAX = 0.4;
+    public final double DTI_RANGE4_MIN = 0.3;
+    public final double DTI_RANGE4_MAX = 0.36;
+
+    public final double SCORE2 = 0.2;
+    public final double SCORE3 = 0.3;
+    public final double SCORE45 = 0.45;
+    public final double SCORE5 = 0.5;
+    public final double SCORE65 = 0.65;
+    public final double SCORE7 = 0.7;
+    public final double SCORE8 = 0.8;
+    public final double SCORE85 = 0.85;
+    public final double SCORE1 = 1.0;
+
+    public final String VERY_HIGH = "Very High";
+    public final String HIGH = "High";
+    public final String MEDIUM = "Medium";
+    public final String LOW = "Low";
+    public final String VERY_LOW = "Very Low";
+}
